feat: implement recommended react-refresh refactor changes
    - this allows fast refreshing of the Vite server
