+++ .eslintrc.json
@@ -36,7 +36,10 @@
     "quotes": ["warn", "single"],
     "react-hooks/rules-of-hooks": "error",
     "react-hooks/exhaustive-deps": "warn",
-    "react/react-in-jsx-scope": "off",
-    "react-refresh/only-export-components": "off"
+    "react-refresh/only-export-components": [
+      "warn",
+      { "allowConstantExport": true }
+    ],
+    "react/react-in-jsx-scope": "off"
   }
 }
