+++ src/components/Liquidity/LiquiditySelector.tsx
@@ -16,9 +16,9 @@ import {
   formatAmount,
   formatPrice,
   formatMaximumSignificantDecimals,
-  roundToSignificantDigits,
   formatPercentage,
 } from '../../lib/utils/number';
+import { getRangeIndexes } from './helpers';
 import { Token, getTokenId } from '../../lib/web3/utils/tokens';
 import { useOrderedTokenPair } from '../../lib/web3/hooks/useTokenPairs';
 import { useTokenPairTickLiquidity } from '../../lib/web3/hooks/useTickLiquidity';
@@ -1061,28 +1061,6 @@ function getRangePositions(
   return [
     rangeMinIndex + (rangeMinIndex <= roundedCurrentPriceIndex ? -1 : 0),
     rangeMaxIndex + (rangeMaxIndex >= roundedCurrentPriceIndex ? +1 : 0),
-  ];
-}
-
-export function getRangeIndexes(
-  currentPriceIndex: number | undefined,
-  fractionalRangeMinIndex: number,
-  fractionalRangeMaxIndex: number
-) {
-  const roundedCurrentPriceIndex =
-    currentPriceIndex && Math.round(currentPriceIndex);
-  const rangeMinIndex = roundToSignificantDigits(fractionalRangeMinIndex);
-  const rangeMaxIndex = roundToSignificantDigits(fractionalRangeMaxIndex);
-  // align fractional index positions to whole tick index positions
-  // for the min and max cases
-  if (roundedCurrentPriceIndex === undefined) {
-    return [rangeMinIndex - 0.5, rangeMaxIndex + 0.5];
-  }
-  return [
-    Math.ceil(rangeMinIndex) +
-      (rangeMinIndex >= roundedCurrentPriceIndex ? -1 : 0),
-    Math.floor(rangeMaxIndex) +
-      (rangeMaxIndex <= roundedCurrentPriceIndex ? +1 : 0),
   ];
 }
