+++ src/components/cards/LimitOrderContext.tsx
@@ -6,58 +6,10 @@ import {
   useMemo,
   useState,
 } from 'react';
-import { LimitOrderType } from '@duality-labs/dualityjs/types/codegen/duality/dex/tx';
-
-export type LimitOrderTypeKeys = keyof Omit<
-  typeof LimitOrderType,
-  'UNRECOGNIZED'
->;
-
-// copy out enum as an object with checks that should break if the type changes
-export const orderTypeEnum: {
-  GOOD_TIL_CANCELLED: typeof LimitOrderType['GOOD_TIL_CANCELLED'];
-  FILL_OR_KILL: typeof LimitOrderType['FILL_OR_KILL'];
-  IMMEDIATE_OR_CANCEL: typeof LimitOrderType['IMMEDIATE_OR_CANCEL'];
-  JUST_IN_TIME: typeof LimitOrderType['JUST_IN_TIME'];
-  GOOD_TIL_TIME: typeof LimitOrderType['GOOD_TIL_TIME'];
-  UNRECOGNIZED: typeof LimitOrderType['UNRECOGNIZED'];
-} = {
-  GOOD_TIL_CANCELLED: 0,
-  FILL_OR_KILL: 1,
-  IMMEDIATE_OR_CANCEL: 2,
-  JUST_IN_TIME: 3,
-  GOOD_TIL_TIME: 4,
-  UNRECOGNIZED: -1,
-};
-
-export const orderTypeTextMap: {
-  [key in LimitOrderTypeKeys]: string;
-} = {
-  IMMEDIATE_OR_CANCEL: 'Immediate Or Cancel',
-  FILL_OR_KILL: 'Fill Or Kill',
-  GOOD_TIL_CANCELLED: 'Good Til Canceled',
-  GOOD_TIL_TIME: 'Good Til Time',
-  JUST_IN_TIME: 'Just In Time',
-};
-export const timePeriods = [
-  'seconds',
-  'minutes',
-  'hours',
-  'days',
-  'weeks',
-] as const;
-export type TimePeriod = typeof timePeriods[number];
-export const timePeriodLabels: {
-  [timePeriod in TimePeriod]: string;
-} = {
-  seconds: 'Seconds',
-  minutes: 'Minutes',
-  hours: 'Hours',
-  days: 'Days',
-  weeks: 'Weeks',
-};
-
-export const defaultExecutionType: LimitOrderTypeKeys = 'FILL_OR_KILL';
+import {
+  AllowedLimitOrderTypeKey,
+  TimePeriod,
+} from '../../lib/web3/utils/limitOrders';
 
 interface FormState {
   amount: string;
@@ -65,7 +17,7 @@ interface FormState {
   triggerPrice: string;
   timeAmount: string;
   timePeriod: TimePeriod;
-  execution: LimitOrderTypeKeys;
+  execution: AllowedLimitOrderTypeKey;
   slippage: string;
 }
 interface FormSetState {
@@ -84,8 +36,10 @@ export const LimitOrderFormSetContext = createContext<Partial<FormSetState>>(
 );
 
 export function LimitOrderContextProvider({
+  defaultExecutionType,
   children,
 }: {
+  defaultExecutionType: AllowedLimitOrderTypeKey;
   children: ReactNode;
 }) {
   const [amount, setAmount] = useState('');
