+++ src/index.tsx
@@ -3,13 +3,11 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import './index.scss';
 
-// add Buffer.from support for '@duality-labs/dualityjs'
-import { Buffer } from 'buffer';
-window.global = window.globalThis;
-window.Buffer = Buffer;
+// add fixes for globals before any of the app and dependencies are loaded
+import './global-fixes';
 
-// ensure App is loaded after Buffer because Keplr needs it on import
-const App = React.lazy(() => import('./pages/App'));
+// ensure App is loaded after Buffer fix because Keplr needs it on import
+import App from './pages/App';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
